requirements.txt
slack-bolt==1.23.0
schedule==1.2.2
python-dotenv==1.1.1
